requirements.txt
streamlit
Pillow
azure.cognitiveservices.vision.computervision ==0.9.0
googletrans==4.0.0-rc1